Fixed bug #785 - Paint tool freezes on right mouse input
From Alan Morris:

We've noticed a bug with the mouse buttons for the paint brush tool.
If you hold the left mouse button, move the mouse to do some painting, and
while still holding the left mouse button, clicked the right mouse button just
once, the paint cursor will freeze in that spot and stay frozen until the mouse
leaves the viewer.  No amount of clicking will continue painting.

We've reproduced it on both mac and windows.