Fixed serious bug causing corruption of data and mask layer id management.  To reproduce:
1) start a new project, import a data layer
2) create mask layer and open paint brush tool
3) save project
4) paint a large letter "A"
5) save project
6) paint a large letter "B"
7) save project
6) load first "New Session" (should be the 2nd from the bottom)
   at this point, you will be back to the blank mask
7) paint a large letter "C"
8) save project
9) paint a large letter "D" 
10) save project
11) reload latest session, or reload the project

Now you will have large letter "A" instead of "C" and "D".  If you go
through each session you will find that the "D" is nowhere to be
found, that data has been lost.

The problem is that when a session is loaded from the current project,
the generation count is reset to 0, and then only incremented as large
as the largest number from the current session.  Hence, if you load an
older session, your new layer changes will be lost as those layer
generation numbers already existed.

To fix it, whenever a session is loaded (not just a project), we now
reload the generation number from the project to the datablock
manager.

It appears that the provenance system has the same bug, so I have
added the same fix for it, though since I do not use the provenance
system, I have not tested it.