Fixed bug in isosurface generation.
If the number of faces is over 1,000,000 on a Z-iteration, it is
broken into multiple parts, but if there are no more faces after the
split, then another, empty part, is added, which later causes a crash
(at least on windows) when uploading the vertex buffer.  This fix
prevents the creation of the empty part.